Consolidated density-based clustering examples
